requirements.txt
python-fasthtml
langchain
langchain-openai
python-dotenv